Add files via upload
This is a compiled firmware.hex that "should" work for the Sovol SV03 since it appears to use the Sovol SV01 silent board.  Everything is the same as CopterTec's SV01 Silent Board with BLTouch except the bed sized has been edited in Configuration.h to 350x350x400.  It seems to be working for me.  Use at your own risk!